requirements.txt
# Core dependencies
python-dotenv==1.0.0
supabase==2.0.3
asyncpg==0.28.0

# Agent frameworks
langgraph==0.0.19
langchain==0.0.335
langchain-openai==0.0.2.post1

# WhatsApp integration
requests==2.31.0

# Testing
pytest==7.4.0

# Utilities
pydantic==2.4.2
tqdm==4.66.1